+++ src/bin_wasm_debugger/messages.ml
@@ -43,9 +43,14 @@ let parsed_string_encoding =
    alternative encoding for {Sc_rollup_inbox_message_repr.t} that only encodes
    `Internal Transfer` and `External`. In the case of `Internal Transfer`, only
    the Micheline payload is mandatory, the other field are taken from the
-   default one if they are missing. *)
+   default one if they are missing.
+
+   It can also take already serialized messages, if the input does not belong
+   to the two cases above. This allows to inject arbitrary messages in
+   the rollup. *)
 let input_encoding default_sender default_source default_destination :
-    [< `Inbox_message of Sc_rollup.Inbox_message.t] Data_encoding.encoding =
+    [< `Inbox_message of Sc_rollup.Inbox_message.t | `Serialized of string]
+    Data_encoding.encoding =
   let open Data_encoding in
   union
     [
@@ -84,6 +89,17 @@ let input_encoding default_sender default_source default_destination :
         (fun msg ->
           `Inbox_message
             (External (Hex.to_string (`Hex msg) |> Option.value ~default:"")));
+      case
+        (Tag 2)
+        ~title:"Serialized"
+        (obj1 (req "serialized" string))
+        (function
+          | `Serialized msg ->
+              let (`Hex msg) = Hex.of_string msg in
+              Some msg
+          | _ -> None)
+        (fun msg ->
+          `Serialized (Hex.to_string (`Hex msg) |> Option.value ~default:""));
     ]
 
 (* Represent a set of inboxes, i.e. a set of set of inputs. The position of an
@@ -116,7 +132,8 @@ let parse_inboxes inputs Config.{sender; source; destination; _} =
                   Protocol.Alpha_context.Sc_rollup.Inbox_message.(
                     serialize input
                     |> Result.map unsafe_to_string
-                    |> Environment.wrap_tzresult |> Lwt.return))
+                    |> Environment.wrap_tzresult |> Lwt.return)
+              | `Serialized input -> Lwt.return_ok input)
             inputs)
         full_inputs
   | Error e -> Error_monad.failwith "%s" e